--- index.js
@@ -1,5 +1,0 @@
-import createApp from './app.js';
-
-const app = createApp();
-
-export default app;
