+++ index.ts
@@ -1,0 +1,5 @@
+// This file serves as the main entry point for the application
+import createApp from '#src/app.js';
+
+// Execute the createApp function to create and start the server
+createApp();
