+++ middleware/setupCsrf.ts
@@ -1,4 +1,7 @@
 import { csrfSync } from "csrf-sync";
+import { Application, Request, Response, NextFunction } from 'express';
+// Import CSRF token type definitions
+import '../types/csrf-types.js';
 
 /**
  * Sets up CSRF protection for an Express application.
@@ -6,17 +9,18 @@ import { csrfSync } from "csrf-sync";
  * - Protects against CSRF attacks using `csrfSync`.
  * - Ensures CSRF tokens are available in views for forms.
  *
- * @param {import('express').Application} app - The Express application instance.
+ * @param {Application} app - The Express application instance.
  */
-export const setupCsrf = (app) => {
-    const {csrfSynchronisedProtection} = csrfSync({
+export const setupCsrf = (app: Application): void => {
+    const { csrfSynchronisedProtection } = csrfSync({
         /**
          * Extracts the CSRF token from the request body.
          *
-         * @param {import('express').Request} req - The incoming request object.
+         * @param {Request} req - The incoming request object.
          * @returns {string|undefined} The CSRF token if present, otherwise undefined.
          */
-        getTokenFromRequest: (req) => req.body._csrf,
+        getTokenFromRequest: (req: Request): string | undefined => 
+            req.body && req.body._csrf ? req.body._csrf : undefined,
     });
 
     /**
@@ -29,11 +33,11 @@ export const setupCsrf = (app) => {
      * Middleware to expose the CSRF token to views.
      * Adds `res.locals.csrfToken`, making it accessible in templates.
      *
-     * @param {import('express').Request} req - The incoming request object.
-     * @param {import('express').Response} res - The response object.
-     * @param {import('express').NextFunction} next - Callback to pass control to the next middleware.
+     * @param {Request} req - The incoming request object.
+     * @param {Response} res - The response object.
+     * @param {NextFunction} next - Callback to pass control to the next middleware.
      */
-    app.use((req, res, next) => {
+    app.use((req: Request, res: Response, next: NextFunction): void => {
         if (typeof req.csrfToken === "function") {
             res.locals.csrfToken = req.csrfToken(); // Makes CSRF token available in views
         }
