+++ types/sass-plugin-types.ts
@@ -1,0 +1,7 @@
+// SassPlugin type definitions
+
+export interface SassPluginOptions {
+	resolveDir?: string;
+	transform?: (source: string) => string;
+	// Add other possible options
+}
