+++ utils/buildHelper.ts
@@ -1,11 +1,10 @@
-// src/utils/buildHelper.ts
 import fs from 'fs';
 
 /**
  * Generate a random build number as a string.
  * @returns {string} - A random build number.
  */
-export const getBuildNumber = () => {
+export const getBuildNumber = (): string => {
   return Math.floor(Math.random() * 10000).toString();
 };
 
@@ -16,7 +15,7 @@ export const getBuildNumber = () => {
  * @param {string} extension - The extension of the build files.
  * @returns {string} - The name of the latest build file or an empty string if none found.
  */
-export const getLatestBuildFile = (directory, prefix, extension) => {
+export const getLatestBuildFile = (directory: string, prefix: string, extension: string): string => {
   const files = fs.readdirSync(directory);
   const pattern = new RegExp(`^${prefix}\\.\\d+\\.${extension}$`);
   const matchingFiles = files.filter(file => pattern.test(file));
